ubocc/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-dotenv==1.0.1
google-generativeai==0.5.4
pandas==2.2.2
pydantic==2.7.1
python-multipart==0.0.9
httpx==0.27.0
aiosqlite==0.20.0
